Moved same functions to gofmt/utils.go and fix/utils.go
